Complete Overhaul Of All Code.
Made a Iridesent 3D Prism for the Prototype Logo.
The Additional Buttons now prints that it works to the main screen.
God Roles toggles now work, & can be used together.

Added a "Go to Developers Section"
Which includes that I (Pretisy) made the program with Dear IMGUI as a DirectX12 Application

Added a "Clear Display Content"
To clear Displayed content on the Main Screen.

All with 50% more Code!